+++ flutter/example/lib/join_route.dart
@@ -66,7 +66,7 @@ class _JoinRouteState extends State<JoinRoute> {
               style: TextButton.styleFrom(
                 foregroundColor: Colors.white,
                 backgroundColor: Colors.blue,
-                disabledForegroundColor: Colors.grey.withOpacity(0.38),
+                disabledForegroundColor: Colors.grey.withAlpha(0x60),
               ),
               onPressed: () {
                 Navigator.push(