+++ proto/control-api/src/control/member.rs
@@ -1,11 +1,17 @@
 //! [`Member`] definitions.
 
-use std::{collections::HashMap, fmt, time::Duration};
+use std::{
+    collections::HashMap,
+    fmt::Write as _,
+    hash::{Hash, Hasher},
+    time::Duration,
+};
 
 use derive_more::{AsRef, Display, Error, From, FromStr, Into};
 use ref_cast::RefCast;
+use secrecy::{ExposeSecret as _, SecretString};
 #[cfg(feature = "serde")]
-use serde::{Deserialize, Serialize};
+use serde::{Deserialize, Serialize, Serializer};
 use url::Url;
 
 use super::{endpoint, room, Pipeline};
@@ -143,13 +149,21 @@ pub struct Sid {
     pub creds: Option<PlainCredentials>,
 }
 
-impl Display for Sid {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}/{}/{}", self.public_url, self.room_id, self.member_id)?;
+impl Sid {
+    /// Renders the [URI] string of this [`Sid`]
+    ///
+    /// [URI]: https://en.wikipedia.org/wiki/Uniform_Resource_Identifier
+    // TODO: Return `SecretString` once `secrecy` crate allows to unwrap it:
+    //       https://github.com/iqlusioninc/crates/issues/1182
+    #[must_use]
+    pub fn to_uri_string(&self) -> String {
+        let mut sid =
+            format!("{}/{}/{}", self.public_url, self.room_id, self.member_id);
         if let Some(plain) = &self.creds {
-            write!(f, "?token={plain}")?;
+            write!(sid, "?token={}", plain.expose_str())
+                .expect("writing to `String` never fails");
         }
-        Ok(())
+        sid
     }
 }
 
@@ -278,21 +292,53 @@ impl Credentials {
 }
 
 /// Plain [`Credentials`] returned in a [`Sid`].
-#[derive(
-    AsRef,
-    Clone,
-    Debug,
-    Display,
-    Eq,
-    From,
-    Hash,
-    Into,
-    Ord,
-    PartialEq,
-    PartialOrd,
-)]
-#[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
+#[derive(AsRef, Clone, Debug)]
+#[cfg_attr(feature = "serde", derive(Deserialize))]
 #[cfg_attr(feature = "serde", serde(transparent))]
-#[from(&str, String)]
-#[into(String)]
-pub struct PlainCredentials(Box<str>); // TODO: Use `secrecy` crate.
+pub struct PlainCredentials(SecretString);
+
+impl PlainCredentials {
+    /// Provides access to the underlying secret [`str`].
+    #[must_use]
+    pub fn expose_str(&self) -> &str {
+        self.0.expose_secret()
+    }
+}
+
+impl<T> From<T> for PlainCredentials
+where
+    T: Into<String>,
+{
+    fn from(value: T) -> Self {
+        Self(value.into().into())
+    }
+}
+
+impl Hash for PlainCredentials {
+    fn hash<H: Hasher>(&self, state: &mut H) {
+        self.expose_str().hash(state);
+    }
+}
+
+impl Eq for PlainCredentials {}
+
+impl PartialEq for PlainCredentials {
+    fn eq(&self, other: &Self) -> bool {
+        use subtle::ConstantTimeEq as _;
+
+        self.expose_str()
+            .as_bytes()
+            .ct_eq(other.expose_str().as_bytes())
+            .into()
+    }
+}
+
+#[cfg(feature = "serde")]
+impl Serialize for PlainCredentials {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+    {
+        self.0.expose_secret().serialize(serializer)
+    }
+}