+++ CHANGELOG.md
@@ -6,6 +6,47 @@ All user visible changes to this project will be documented in this file. This p
 
 
 
+## [master] · unreleased
+[master]: /../../tree/master
+
+[Diff](https://github.com/instrumentisto/medea-jason/compare/medea-jason-0.7.0...master)
+
+### Upgraded
+
+- Dependencies:
+    - [`flutter_rust_bridge`] to 2.7.0 version. ([#195])
+
+[#195]: /../../pull/195
+
+
+
+
+## [0.7.0] · 2024-10-29
+[0.7.0]: /../../tree/medea-jason-0.7.0
+
+[Diff](https://github.com/instrumentisto/medea-jason/compare/medea-jason-0.6.1...medea-jason-0.7.0)
+
+### BC Breaks
+
+- [Cargo features]:
+    - Replaced `wee_alloc` with `talc`. ([#187])
+
+### Changed
+
+- Dependencies:
+    - Replaced [`wee_alloc`] allocator with [`talc`] for web. ([#187])
+
+### Upgraded
+
+- Dependencies:
+    - [`flutter_rust_bridge`] to 2.4.0 version. ([#188])
+
+[#187]: /../../pull/187
+[#188]: /../../pull/188
+
+
+
+
 ## [0.6.1] · 2024-09-11
 [0.6.1]: /../../tree/medea-jason-0.6.1
 
@@ -397,9 +438,12 @@ All user visible changes to this project will be documented in this file. This p
 
 
 
+[Cargo features]: https://doc.rust-lang.org/cargo/reference/features.html
 [MSRV]: https://doc.rust-lang.org/cargo/reference/manifest.html#the-rust-version-field
 [SFU]: https://webrtcglossary.com/sfu
 [Semantic Versioning 2.0.0]: https://semver.org
 [VP9]: https://bloggeek.me/webrtcglossary/vp9
 [`derive_more`]: https://docs.rs/derive_more
 [`flutter_rust_bridge`]: https://docs.rs/flutter_rust_bridge
+[`talc`]: https://docs.rs/talc
+[`wee_alloc`]: https://docs.rs/wee_alloc