+++ flutter/CHANGELOG.md
@@ -6,6 +6,36 @@ All user visible changes to this project will be documented in this file. This p
 
 
 
+## [master] · unreleased
+[master]: /../../tree/master/flutter
+
+See also [`medea-jason` crate `master` changes](/../../tree/master/CHANGELOG.md).
+
+### Upgraded
+
+- Dependencies:
+    - [`flutter_rust_bridge`] to 2.7.0 version. ([#195])
+
+[#195]: /../../pull/195
+
+
+
+
+## [0.7.0] · 2024-10-29
+[0.7.0]: /../../tree/medea-jason-0.7.0/flutter
+
+See also [`medea-jason` crate 0.7.0 changes](/../../tree/medea-jason-0.7.0/CHANGELOG.md).
+
+### Upgraded
+
+- Dependencies:
+    - [`flutter_rust_bridge`] to 2.4.0 version. ([#188])
+
+[#188]: /../../pull/188
+
+
+
+
 ## [0.6.1] · 2024-09-11
 [0.6.1]: /../../tree/medea-jason-0.6.1/flutter